server/requirements.txt
autopep8==1.5.7
click==7.1.2
dnspython==2.1.0
Flask==1.1.2
Flask-Cors==3.0.10
greenlet==1.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mongoengine==0.23.1
pycodestyle==2.7.0
PyJWT==2.1.0
pymongo==3.11.4
PyMySQL==1.0.2
python-dotenv==0.17.0
six==1.15.0
SQLAlchemy==1.4.11
toml==0.10.2
Werkzeug==1.0.1
